XVIZValidator, check JSON correctness
This provides a simple API that validates provided JSON objects are
the types you would expect.  It has a generic API that lets you check
any subobject of the spec and a set of high level methods to validate
the types that the builder objects are most concerned about.
