+++ modules/schema/examples/core/stream_set/complete.json
@@ -1,5 +1,27 @@
 {
   "timestamp": 1001.3,
+  "poses": {
+    "/vehicle_pose": {
+      "timestamp": 1001.3,
+      "mapOrigin": {
+        "longitude": 40.4641969,
+        "latitude": -79.9733218,
+        "altitude": 219
+      },
+      "position": [238.4, 1002.5, 5.0],
+      "orientation": [0.5, 10.0, 84.5]
+    },
+    "/original/vehicle_pose": {
+      "timestamp": 1001.3,
+      "mapOrigin": {
+        "longitude": 40.46423,
+        "latitude": -79.97384,
+        "altitude": 220
+      },
+      "position": [123.4, 578.1, 3.5],
+      "orientation": [1.5, 95.0, 24.5]
+    }
+  },
   "primitives": {
     "/object/points": {
       "primitives": [
