+++ test/modules/schema/index.js
@@ -1,3 +1,5 @@
 // Set up a configuration (TODO/OSS - this should be a neutral config)
 import './validator.spec';
 import './examples.spec';
+import './file-validation.spec.js';
+import './data.spec.js';
